Refactor commands
