+++ src-tauri/src/commands/mod.rs
@@ -1,0 +1,8 @@
+mod bank_account_transactions;
+mod bank_accounts;
+mod bank_providers;
+
+// re-exports of commands
+pub use bank_account_transactions::*;
+pub use bank_accounts::*;
+pub use bank_providers::*;
